PREVIEW: Add option to list snapshots from all configs
$ snapper list -a

will print the same output as 'list' but each table is preceded with
config name and subvolume. This is convenient way to see all configured
snapshots in the system, compared to listing all configs and then typing
each command separately.

Note: the implementation is dumb, patch lacks documentation but works
well enough to demonstrate the idea.