btrfs: do not wrongly report failure when doing btrfs comparison
snapper would always fail using btrfs features to perform a comparison
between two read-only snapshots, and fallback to calling stat on each
file.

The reason was that btrfs_read_and_process_send_stream would first
return 0, because it successfully did its work, and the second time
it would return -ENODATA, because no more data was present in the
stream to be processed.

We now check that we fail due to ENODATA only if it was the first time
we tried to call btrfs_read_and_process_send_stream.

In my case, a comparison took initially ~30s, and came down to ~300ms
with this fix.