Ensure snapperd runs with 'idle' IO and CPU scheduling (boo#1099494)
With https://github.com/openSUSE/snapper/pull/391 we already asked the
background "cleanup" service to take idle IO and CPU priority so for
sake of consistency and potentially fix bad responsiveness of systems
the snapperd service which is triggered over dbus should be asked to run
in a 'nice' way.

This change is motivated by an observation I made on my own system: The
timeline service was triggered at around the same time a personal sync
and backup script was triggered. Then the system turned unresponsive for
some minutes. I guess asking the snapperd service to act with idle
priority might help a bit in similar situations.

Changing the the timeline or cleanup service does not help as they would
only trigger snapperd and not produce significant IO themselves and the
nice levels are not communicated to snapperd otherwise.

I have been running this change for some days now and did not observe
any problems. I could ensure that snapperd would still start up and
trigger the necessary actions and I have not observed any negative
impact so it might fix the original issue but this is not ensured.