Add option to reuse existing subvolume
If a subvolume exists already mksubvolume will abort. In certain
situations however, e.g. if there was a rollback to a previous snapshot,
it may be desirable to be able to reuse this subvolume.

This will make it possible to use the mksubvolume command in package
scripts.

To avoid having to duplicate the exception handler code multiple times
a C wrapper object was introduced to automatically clean up in the
destructor.

Fixes https://github.com/openSUSE/snapper/issues/236.