Run the tests twice, with and without validation.
Why make it so complicated?

Before, coverage measurement gets messed up (the numbers go down and I
look, and feel, like a fool). It's because forwarding-zypp-plugin and
snapper-zypp-plugin run at the same time and both want to write to the
same GCOV data files. The GCOV instrumentation will merge the data, but
only if the programs run in sequence. Being run in parallel, the simpler
plugin somehow wins, having covered much less code :-(

(
There is a way to tell GCOV to write to a different location, with
GCOV_PREFIX and GCOV_PREFIX_STRIP, but that seems less workable than
this.
https://gcc.gnu.org/onlinedocs/gcc/Cross-profiling.html#Cross-profiling
)